+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v0.8.0
+- Support browser.
+- Support extra header.
+- Add copy method to VGMCommand.
+- Subclassing VGMWaitCommand.
+- Decompress .vgz file internally.
+
 # v0.6.1
 - Fix the bug where SN76489 is recognized as T6W28.